refactor and clean
