+++ reconstruction/rtpc/src/main/java/org/jlab/rec/rtpc/hit/TrackFinder.java
@@ -1,0 +1,141 @@
+//Author: David Payette
+
+/* This code sorts pad signals which have been integrated into 120 ns time slices into tracks
+ * based on their relative positions in space, and how close in time the signals occur
+ */
+
+package org.jlab.rec.rtpc.hit;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+
+public class TrackFinder {
+    
+    private TrackUtils tutil = new TrackUtils();
+    private TrackMap TIDMap = new TrackMap();
+    private List<Integer> TIDList;
+    private Track track;
+    private HashMap<Integer, double[]> ADCMap;
+    private List<Integer> PadNum;
+    private int TrigWindSize;
+    private int StepSize = 120;//Bin Size of Dream Electronics Output
+    private double adcthresh = 5e-4; 
+    private int padloopsize;// = PadNum.size();
+    private boolean padSorted = false; 
+    private List<Integer> padTIDlist = new ArrayList<>();
+    private List<Integer> padlist; 
+    private int pad = 0;
+    private double adc = 0; 
+    private int timeadjlimit = 4; 
+    private int parenttid = -1;
+    private String method = "phiz";
+    private int minhitcount = 5; 
+
+    public TrackFinder(HitParameters params) {
+        /*	
+         *Initializations 
+         */
+        
+        ADCMap = params.get_R_adc();
+        PadNum = params.get_PadNum();
+        TrigWindSize = params.get_TrigWindSize();
+        padloopsize = PadNum.size();
+        /*
+         * Main Algorithm
+         */
+        TIMELOOP: //Loop over all times
+        for(int time = 0; time < TrigWindSize; time += StepSize) { //Steps of 120 up to TrigWindSize = 10000
+
+            PADLOOP: //Loop over all pads
+            for(int padindex = 0; padindex < padloopsize; padindex++) {
+                padSorted = false;  //Flag to be set when the pad is assigned to a track
+                padTIDlist.clear(); //List of all TIDs assigned to the pad starts empty
+                pad = PadNum.get(padindex);	
+                adc = ADCMap.get(pad)[time];
+
+                if(adc > adcthresh) { //pad adc threshold check
+                    PadVector PadVec = params.get_padvector(pad); //initializes the x,y,z,phi for pad
+                    TIDList = TIDMap.getAllTrackIDs(); //Retrieve list of all available TIDs
+
+                    TIDLOOP: //Loop over all Track IDs 
+                    for(int tid : TIDList) {
+                        track = TIDMap.getTrack(tid); //Get track with current tid
+
+                        TIMECHECKLOOP: //Loop over current and former times
+                        for(int timecheck = time; timecheck > 0 && timecheck >= time - timeadjlimit*StepSize; timecheck -= StepSize) {
+                            padlist = track.getTimeSlice(timecheck); //Get pads assigned to current time slice
+                            if(!padlist.contains(pad)) { //Ensures pad isn't already assigned here
+                                PADCHECKLOOP: //Loop over pads 
+                                for(int checkpad : padlist) {		
+                                    PadVector checkpadvec = params.get_padvector(checkpad);	
+                                    if(tutil.comparePads(PadVec, checkpadvec, method)) { //compares the position of two pads
+                                        track.addPad(time, pad);			//assign pad to track
+                                        padSorted = true;				//flag set
+                                        padTIDlist.add(tid);				//track the TID assigned
+                                        break TIMECHECKLOOP;                            //no need to continue checking previous times
+                                    } //END PAD COMPARE
+
+                                } //END PADCHECKLOOP
+
+                            } else {//pad is somehow in this time slice already so lets go ahead and add it to the current time slice
+                                    track.addPad(time, pad);
+                                    padSorted = true;
+                                    padTIDlist.add(tid);
+                                    break TIMECHECKLOOP;
+                            }
+
+                        } //END TIMECHECKLOOP 
+
+                    } //END TIDLOOP 
+
+                    if(!padSorted) { //we need a new TID if we get here, the pad was never assigned an ID
+                            TIDMap.addTrack(new Track(time,pad));
+                    }
+
+                    if(padTIDlist.size()>1) { //if a pad gets more than 1 ID let's merge the IDs	
+                        for(int tidtemp : padTIDlist) {
+                            if(tidtemp == padTIDlist.get(0)) {
+                                parenttid = padTIDlist.get(0);
+                            } else {
+                                TIDMap.mergeTracks(parenttid, tidtemp); 
+                            }
+                        }						
+                    }
+
+                } //END ADC THRESH CHECK
+
+            } //END PADLOOP
+
+        } //END TIMELOOP
+
+        //END MAIN ALGORITHM
+
+        /*
+         * Clean up and flag tracks
+         */
+
+        for(int tid : TIDMap.getAllTrackIDs()) { //We need to remove tracks with not enough pads to save time later
+            Track tempt = TIDMap.getTrack(tid);
+            if(tempt.uniquePadCountTotal() < minhitcount) {
+                    TIDMap.removeTrack(tid);
+            }
+        }
+
+        //System.out.println("This event has " + TIDMap.getAllTrackIDs().size() + " tracks");
+
+        //TODO Flag crossing tracks; for now flag all tracks
+        for(int tid : TIDMap.getAllTrackIDs()) {
+            Track t = TIDMap.getTrack(tid);
+            t.flagTrack();
+        }
+
+        /*
+         * Output
+         */
+
+        params.set_trackmap(TIDMap);
+
+    }       
+	
+}
